requirements.txt
torch>=2.2.0
transformers>=4.45.0
accelerate>=0.29.0
qwen-vl-utils>=0.0.2
gradio>=4.21.0
pillow>=10.2.0
bitsandbytes>=0.43.0
